Implement level deboosting
fish delevel if the catch has failed and they haven't been leveled

modified the catch functions to map a customproperty indicating the fish is caught. This is needed for checks because the function that triggers deboosting is called multiple times in fishing float procesing function.

Additionally 2 new patches, one that removes that property when the fish is dropped, and the deboosting itself.