Remove DropItem() patch
Don't boost fish that were dropped from inventory at all anymore.

Removed said patch because there was an issue with stacks in inventory. Stacked items in inventory do not have their object autonomy, instead one of the objects gets an m_stack attribute incremented, therefore we can't know which fish was which in order to boost/deboost it after it drops. Therefore fish are now locked.